requirements.txt
PyGithub>=2.3.0
PyYAML>=6.0.1
requests>=2.31.0
pytest>=7.4.0
